refactor(executor): move 8 utility functions to streaming.go (GH-1140)
Move small standalone functions with no cross-dependencies:
- estimateCost(): token usage cost calculation
- extractCommitSHA(): parse git commit SHA from output
- isValidSHA(): validate git SHA format
- formatToolMessage(): format tool usage for display
- truncateText(): text truncation with ellipsis
- min(): return smaller of two ints
- emitAlertEvent(): send alert events
- dispatchWebhook(): dispatch webhook events

Add parseStreamEvent() wrapper method for test compatibility.
Remove functions from runner.go, add required imports to streaming.go.
Verify with 'go build' - all compilation successful.
