+++ .agent/tasks/gh-1140.md
@@ -1,0 +1,10 @@
+# GH-1140
+
+**Created:** 2026-02-14
+
+## Problem
+
+Move `estimateCost()`, `extractCommitSHA()`, `isValidSHA()`, `formatToolMessage()`, `truncateText()`, `min()`, `emitAlertEvent()`, `dispatchWebhook()`. Small standalone functions with no cross-dependencies. Verify with `go build`.
+
+## Acceptance Criteria
+
